refactor: authKeys를 keys.ts 리프 모듈로 분리
- mutation → query 결합을 끊고 album/api/keys.ts 패턴과 동일
