+++ apps/web/src/features/auth/api/queries.ts
@@ -3,10 +3,7 @@ import { useQuery } from '@tanstack/react-query';
 import type { User } from '@/entities/user';
 import { http } from '@/shared/api';
 
-export const authKeys = {
-  all: ['auth'] as const,
-  me: () => [...authKeys.all, 'me'] as const,
-};
+import { authKeys } from './keys';
 
 export function useMe() {
   return useQuery({
